% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_overlap.R
\name{find_overlap}
\alias{find_overlap}
\title{Find overlap between vessel track points}
\usage{
find_overlap(
  df,
  time_range,
  speed_range,
  time_col = "DATE",
  speed_col = "SPE",
  lon_col = "LON",
  lat_col = "LAT"
)
}
\arguments{
\item{df}{A data frame consisting of a time series. Each observation is a point in time. The time points must be evenly spaced.}

\item{time_range}{A time range, in the form of a two-element vector.}

\item{speed_range}{A vessel speed range, also in the form of a two-element vector.}

\item{time_col}{A string indicating the name of the column that contains the time values.}

\item{speed_col}{A string indicating the name of the column that contains the speed values.}

\item{lon_col}{A string indicating the name of the column that contains the longitude values.}

\item{lat_col}{A string indicating the name of the column that contains the latitude values.}
}
\value{
A \link[tibble:tibble-package]{tibble} whose observations are the closest
points, i.e., the first row contains information about the point closest to
the first, the second row, about the point closest to the second, and so
on. These are the columns:
\describe{
\item{\code{arg_min_dist}}{The index of closest point.}
\item{\code{min_dist}}{The distance to the closest point.}
}
}
\description{
Find overlap between vessel track points.
\strong{Note: this function is highly experimental and still needs to be
documented.}
}
