% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishy-package.R
\docType{package}
\name{fishy-package}
\alias{fishy}
\alias{fishy-package}
\title{fishy: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\author{
\strong{Maintainer}: Ramiro Magno \email{ramiro.magno@gmail.com} (\href{https://orcid.org/0000-0001-5226-3441}{ORCID})

Other contributors:
\itemize{
  \item CINTESIS [copyright holder, funder]
}

}
\keyword{internal}
