% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect2.R
\name{intersect2}
\alias{intersect2}
\title{Intersect a sequence and a vector}
\usage{
intersect2(range, x)
}
\arguments{
\item{range}{A two-element integer vector.}

\item{x}{An integer vector.}
}
\value{
An integer vector.
}
\description{
This function intersects a sequence, i.e. an sequential integer vector with
an integer vector \code{x}. The sequence \code{range} is provided as an interval in the
form of a two element vector with start and end indices.
}
\examples{
intersect2(range = c(3, 8), x = c(1, 5, 10))

}
