% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect3.R
\name{intersect3}
\alias{intersect3}
\title{Intersect a sequence and a logical vector}
\usage{
intersect3(range, lgl)
}
\arguments{
\item{range}{A two-element integer vector.}

\item{lgl}{A logical vector.}
}
\value{
An integer vector.
}
\description{
This function intersects a sequence, i.e. an sequential integer vector with a
logical vector \code{lgl}. The sequence \code{range} is provided as an interval in the
form of a two element vector with start and end indices.
}
\examples{
intersect3(range = c(2, 4), lgl = c(FALSE, TRUE, FALSE, TRUE, TRUE))

}
