% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spDistsN1.R
\name{spDistsN1}
\alias{spDistsN1}
\title{Euclidean or Great Circle distance between points}
\usage{
spDistsN1(pt_x, pt_y, pts_x, pts_y, longlat = TRUE)
}
\arguments{
\item{pt_x}{A \code{double} scalar, the longitude of the single point.}

\item{pt_y}{A \code{double} scalar, the latitude of the single point.}

\item{pts_x}{A \code{double} vector, the longitude of the set of the other points.}

\item{pts_y}{A \code{double} vector, the latitude of the set of the other points.}

\item{longlat}{A \code{logical} scalar indicating whether to calculate the Great
Circle (WGS84 ellipsoid) distance (\code{TRUE}) or the Euclidean distance
(\code{FALSE}).}
}
\value{
A \code{double} vector of distances in the original metric of the points.
}
\description{
The function returns a vector of distances between a single 2D point
(\code{pts_x}, \code{pts_y}) and a set of 2D points, provided as two \code{double}
vectors, \code{pts_x} and \code{pts_y}.

The distance calculation can be either the Great Circle distance  (WGS84
ellipsoid, \code{longlat = TRUE}), or the Euclidean distance (\code{longlat = FALSE}).
}
\details{
This function is a thin R wrapper around the C function \code{fishy_dists}. This
function performs no checking on the input, and arguments are passed down
to the C implementation as is. This means you should not use this function
if you do not know what you are doing.
}
\note{
The code is based on the source of \link[sp:spDistsN1]{sp::spDistsN1}. We adapted and
included \link[sp:spDistsN1]{sp::spDistsN1} in this package because of two reasons: (i) to
avoid a dependency to sp, as this is the only function used from that
package; and (ii) to adapt its prototype to accept vectors instead of
matrices, as that fitted better in our workflow.
}
\keyword{internal}
