% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_ranges.R
\name{window_ranges}
\alias{window_ranges}
\title{Determine window ranges}
\usage{
window_ranges(n, start, window = 1, na_rm = FALSE)
}
\arguments{
\item{n}{Number of windows.}

\item{start}{Starting position of the first window.}

\item{window}{Window size.}

\item{na_rm}{A logical indicating whether to remove invalid ranges (i.e.,
whether to keep \code{from} and \code{to} as \code{NA}) or otherwise.}
}
\value{
A matrix of ranges (\code{from} and \code{to}), each row pertains a window.
}
\description{
This function determines ranges (i.e., from and to positions) for windows of
size \code{window}. \code{n} specifies the number of windows.
}
\examples{
window_ranges(10, start = 2, window = 3)

window_ranges(10, start = 2, window = 3, na_rm = TRUE)
}
